ocr_service/requirements.txt
flask==3.0.0
pytesseract==0.3.10
pillow==10.1.0